requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Dance==7.1.0
SQLAlchemy==2.0.41
Werkzeug==3.1.3
python-dotenv==1.0.0
email-validator==2.2.0
oauthlib==3.3.1
PyJWT==2.10.1
Pillow==11.3.0
boto3==1.39.9
botocore==1.39.9
psycopg2-binary==2.9.10
gunicorn==23.0.0 